Tests/LibWeb/Text/expected/HTML/HTMLObjectElement-adoption-does-not-leak-load-delayer.txt

iframe load event
window load event
